fix(element-template-generator): new util for linked resources and test fix
